pyproject.toml
[project]
name = "memory-index"
version = "0.1.0"
description = "Local MCP server for persistent entity/observation/relation memory via semantic search"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]",
    "chromadb",
    "sentence-transformers",
    "transformers",
    "einops",
    "onnxscript",
    "networkx",
]

# NOTE: onnxruntime is NOT listed in dependencies. The correct GPU-specific variant
# (onnxruntime-gpu for NVIDIA, onnxruntime-directml for AMD/Intel, or plain
# onnxruntime for CPU) is installed by setup.bat via scripts/detect_gpu.py.
# These packages share the same Python namespace and CONFLICT if co-installed.
# Never add any onnxruntime variant here — setup.bat handles it.

[project.optional-dependencies]
dev = ["pytest"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
